CLIENT-4692 Remove ChainableQueryBuilder bins() method. Use readingOnlyBins() instead.
Also, add copyright headers to benchmarks code.

+++ benchmarks/src/com/aerospike/benchmarks/LatencyManagerAerospike.java
@@ -1,10 +1,25 @@
+/*
+ * Copyright 2012-2026 Aerospike, Inc.
+ *
+ * Portions may be licensed to Aerospike, Inc. under one or more contributor
+ * license agreements WHICH ARE COMPATIBLE WITH THE APACHE LICENSE, VERSION 2.0.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package com.aerospike.benchmarks;
 
 import java.io.PrintStream;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.aerospike.benchmarks.Constants.*;
-
 
 public class LatencyManagerAerospike implements LatencyManager {
 
@@ -121,7 +136,7 @@ public class LatencyManagerAerospike implements LatencyManager {
         long percent = 0;
 
         if (sum > 0 && value > 0) {
-            percent = Math.round((double) value * 100.0 / sum);
+            percent = Math.round(value * 100.0 / sum);
         }
         String percentString = Long.toString(percent) + "%";
         int spaces = Integer.toString(limit).length() + 4 - percentString.length();
